% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDM2FFSuper.R
\name{SDM2FFSuper}
\alias{SDM2FFSuper}
\title{2-Span Slope Deflection Method(Fixed-Fixed) with Multiple Load}
\usage{
SDM2FFSuper(InputMatrix)
}
\arguments{
\item{InputMatrix}{The matrix with parameter values for SDMFF().}
}
\value{
Moments and Rotations
}
\description{
This function computes the moments and the rotations using slope deflection method(SDM).
}
\details{
Given the span length, magnitude of load, type of load, point of action, Young's Modulus, and interia, the function computes the moments and the rotations using SDM.
}
\note{
The possible values for Typeof1 and Typeof2 are UDL/PL/TDLS/TDLE.\cr
UDL:Uniformly Distributed Load \cr
PL:Point Load\cr
TDLS:Triangular Distributed Load(the maximum value is at the start, see http://www.abzwater.com/sdm)\cr
TDLE:Triangular Distributed Load(the maximum value is at the end, see http://www.abzwater.com/sdm )\cr
}
\examples{
Single load on a span (http://www.abzwater.com/sdm/mload)
inputMatrix=matrix(c(8,8,"PL","PL",40,0,4,1,2,1,3,1), nrow=1, byrow=TRUE)
SDM2FFSuper(inputMatrix)

Multiple load on a span (http://www.abzwater.com/sdm/mload)
inputMatrix=matrix(c(3,2,"PL","PL",40,40,1,1,1,1,1,1,3,2,"PL","PL",40,0,2,1,1,1,1,1,3,2,"PL","UDL",0,5,1,1,1,1,1,1), nrow=3, byrow=TRUE)
SDM2FFSuper(inputMatrix)

}
\references{
(1) Wikipedia contributors. (2018, February 19). Slope deflection method. In Wikipedia, The Free Encyclopedia. Retrieved 09:17, March 26, 2018, from https://en.wikipedia.org/w/index.php?title=Slope_deflection_method&oldid=826540985
\cr
\cr
(2) Norris, Charles Head; John Benson Wilbur; Senol Utku (1976). Elementary Structural Analysis (3rd ed.). McGraw-Hill. pp. 313–326. ISBN 0-07-047256-4.
\cr
\cr
(3) McCormac, Jack C.; Nelson, James K. Jr. (1997). Structural Analysis: A Classical and Matrix Approach (2nd ed.). Addison-Wesley. pp. 430–451. ISBN 0-673-99753-7.
\cr
\cr
}
\seealso{
\link{SDM2FF}
\link{SDM2FS}
\link{SDM2FSC}
}
\author{
\strong{Sivarajah Mylevaganam}
}
\keyword{Civil}
\keyword{Deflection}
\keyword{Engineering}
\keyword{Indeterminate}
\keyword{Slope}
\keyword{Structural}
